Move core tests to new module
